All backend types and interfaces were separated via logic structure
