+++ src/types/backend-interfaces/cv/exportPDFInput.ts
@@ -1,0 +1,11 @@
+export interface IExportPDFInput {
+  html: string;
+  margin?: IMarginInput;
+}
+
+export interface IMarginInput {
+  top: string;
+  bottom: string;
+  left: string;
+  right: string;
+}
