+++ src/types/backend-interfaces/cv/skill.ts
@@ -1,0 +1,20 @@
+import { Mastery } from "@/types/backend-interfaces/skill/mastery";
+
+export interface IAddCVSkillInput {
+  cvId: number;
+  name: string;
+  category?: string;
+  mastery: Mastery;
+}
+
+export interface IUpdateCVSkillInput {
+  cvId: number;
+  name: string;
+  category?: string;
+  mastery: Mastery;
+}
+
+export interface IDeleteCVSkillInput {
+  cvId: number;
+  name: string[];
+}
