+++ src/types/backend-interfaces/deleteResult.ts
@@ -1,0 +1,3 @@
+export interface IDeleteResult {
+  affected: number;
+}
