+++ src/types/backend-interfaces/department.ts
@@ -1,0 +1,18 @@
+export interface IDepartment {
+  id: number;
+  created_at: string;
+  name: string;
+}
+
+export interface ICreateDepartmentInput {
+  name: string;
+}
+
+export interface IUpdateDepartmentInput {
+  departmentId: number;
+  name: string;
+}
+
+export interface IDeleteDepartmentInput {
+  departmentId: number;
+}
