+++ src/types/backend-interfaces/skill/mastery.ts
@@ -1,0 +1,13 @@
+export enum Mastery {
+  Novice,
+  Advanced,
+  Competent,
+  Proficient,
+  Expert,
+}
+
+export interface ISkillMastery {
+  name: string;
+  category?: string;
+  mastery: Mastery;
+}
