+++ src/types/backend-interfaces/user/auth.ts
@@ -1,0 +1,11 @@
+import { IUser } from "@/types/backend-interfaces/user";
+
+export interface IAuthInput {
+  email: string;
+  password: string;
+}
+
+export interface IAuthResult {
+  user: IUser;
+  access_token: string;
+}
