+++ src/types/backend-interfaces/user/avatar.ts
@@ -1,0 +1,10 @@
+export interface IUploadAvatarInput {
+  userId: number;
+  base64: string;
+  size: number;
+  type: string;
+}
+
+export interface IDeleteAvatarInput {
+  userId: number;
+}
