+++ src/types/backend-interfaces/verifyMailInput.ts
@@ -1,0 +1,3 @@
+export interface IVerifyMailInput {
+  otp: string;
+}
